Time stamp quizzes, and make them searchable in admin
All quizzes are now timestamped with their creation time. This makes it
easier to debug error messages, since I can see if the quiz was created
before or after a particular fix.

Quizzes are now displayed in the admin interface, searchable by key.
This also makes it easier to investigate errors.