In quiz mode: Explanation of previous question
Issue #32

When in quiz mode, if a question is answered correctly, offer the user
to view the explanation. Don't expand it by default, in order to not
disturb the user.

Had to remove the fading effect from the "Correct!" h1, as it was now
just confusing.

Does not work with the last question yet.