Add option to retract a question
This will prevent it from being selected as a random question in
training mode, and from being included in a quiz.

If the question is already included in an old quiz, it will however
still be included, and if a user follows a direct link to the question,
it will be displayed.

When a retracted question is shown, a warning is displayed at the top.

The reasons I did it like this, instead of just unpublishing questions I
need to retract, is to avoid breaking old quizzes, and to make it easier
to refer externally to retracted questions.