Allow scheduling questions for publishing in the future
To schedule a question for publishing in the future, set its status to
"Scheduled" and give it a publication time. The "auto_publish"
management command will then publish it when time comes, and post a
tweet about it.

To work, this requires a cron job which runs `manage.py auto_publish`,
as well as a file `.cppquiz-secrets.json` with the following contents:

```
{
    "consumer_key":"",
    "consumer_secret":"",
    "key":"",
    "secret":""
}
```

`consumer_key` and `consumer_secret` comes from the Twitter Developer
Portal, where it's often referred to as "API key" and "API secret key".
`key` and `secret` can be generated using `twurl`.