Improve output in auto_publish.py
It's now silent in the case where there are no questions to publish. My
inbox appreciates this, as all cronjob output is mailed to me.

Error handling is also better. In the case of errors tweeting, an
exception gets thrown by `api.update_status(content)` which we catch,
print and then return 1. Previously I just printed the status, which was
done for debugging while developing, and then I forgot to handle it
properly before putting it into production.