Allow https in tweets
When saving a question with a Tweet text, we check that it contains a
link to the question. Previously the site was served as http, so we
checked for the presence of `http://`. Now we serve the site over https,
so let's support both and check for either `http://` or `https://`.

Fixes #193