Enable timezones
Time zones were never really relevant for this site until we started
auto publishing questions. Then we care about the time actually being
correct. However, this wasn't a problem in practice, since I was the
only one scheduling questions for publishing, and since the server was
in my timezone (I'm on Oslo, server was in Amsterdam). Now that we've
moved to a US based webhost however, it's become an issue in practice.

By enabling `USE_TZ = True`, all datetimes stored in the database will
be stored as UTC, but will be presented to the user in their timezone.

One strange thing that happened here, was that applying migrations
started to fail when timezones were enabled. In particular,
`Question.last_viewed` failed like this:

```
DateTimeField Question.last_viewed received a naive datetime (0001-01-01 00:00:00) while time zone support is active.
(...)
OverflowError: date value out of range
```
First we see a warning that a timezone naive datetime object is being
used, which is because we use a naive datetime object in
`0011_question_last_viewed.py`. It looks like this then gets converted
to UTC, which tries to subtract some hours from it, rendering it out of
range. This all makes sense. What I don't understand however is why the
datetime object is even used in the first place when we're just applying
migrations. There should be no datetime computations being done at this
point. (I expected no such computations to occur until we actually start
running the site/tests, at which point we've also applied
`0018_auto_20201219_2341.py`, which removes this default entirely.)

So I had to hack around this in the old migration. I've tested it both
locally and on the beta site, so I'll commit this even though I don't
fully understand why it's necessary.

Fixes #194