Stop using pinned requirements
Only use pinned versions for the frozen requirements in
requirements.frozen.txt. Use unpinned versions in the main
requirements.txt. This makes it easier to upgrade minor/patch versions.

The parameterized package is intentionally left pinned, since it's not
on a stable version yet and only used in tests anyway.

Also upgrade to the latest versions of dependencies.