Fix bug in formatting standard references
The `?` at the end of the `section_name` regex was a remnant from when
we used both names and numbers to refer to sections in the standard. It
should have been removed in 78dac93482.

When I failed to remove it, `standard_ref` would try to format section
references using numbers, such as `§1.2.3` (see the unit test). We no
longer use those kinds of references, but some text, when copied from
the standard, still does. So we need to just leave those alone.

Without this fix, we'd get an exception in `format_reference` since
`section_name` would be `None`:

```
TypeError: can only concatenate str (not "NoneType") to str
```