Automatically format questions on save
But leave this as optional, since we want to manually format some of the
questions.

For now, since we don't have an advanced editor for non-admin
contributors, leave the default to False. That way, we preserve any
formatting they have done, intentionally or not. Then, admins can turn
it on if they want to during review.

We can change this later if we make a more advanced user interface for
contributors.