Factor session out of UserData
This was a messy way to do it, let's remove the session data from the
UserData class and instead provide free functions to load and save
UserData from the Django session. Let's add some type annotations too
while at it.

Also, let's do our own serialization to and from regular dicts, and not
rely on Django's pickle serializer. That should help with #348.