Factor session out of QuizInProgress
This was a messy way to do it, let's remove the session data from the
QuizInProgress class and instead provide free functions to load and save
QuizInProgress from the Django session.

Also, let's do our own serialization to and from regular dicts, and not
rely on Django's pickle serializer. That should help with #348.